LLM_CLI_FILEACCESS/requirements.txt
python-dotenv==1.0.1
groq==0.9.0
colorlog==6.9.0
coloredlogs==15.0.1
httpx==0.24.1